Changes to be committed: 	new file:   addPurchaseSQ.sql 	modified:   displayPurchaseList.sql 	new file:   getCustomersAll.sql 	new file:   getCustomersByLastName.sql 	modified:   project_DDL.sql 	modified:   project_testData.sql
 Changes not staged for commit:
	modified:   ../php/ArtGalleryMain.php

I've changed the purchase query to "simulate" a check constraint, but I'm not happy with it.  See my email.  I want to go back to the foreign key version. for now, the addPurchaseSQ.sql query is the one to use.  I didn't modify the original non-SQ file, but it wont work on the current schema.  If we go back I'll fix it.
I've added additional initial data to the testdata script, so that there is something in every table, and I've deployed the current set of tables and the test data to the ONID instance of mySQL for out class.  I've added the credentials in ArtGalleryMain.php.
Others of the existing queries needed modification due to the sales change.  You can see them in the change list.  I've created one query to address you main line 477 section and a child query that filters by last name (using like, try some partials, like %G% as your last name to see the effect.)

That's a s far as I got.  I'll continue to work on parent/child queries tomorrow.  Please let me know if you can't connect as early as you possibly can.

Tom